ScratchNotesActivity shows saved recipes, and you can delete them from there using the "x". It will delete it from the savedRecipes list on your profile, not the actual recipe. When clicking on a recipe on the ScratchNotesActivity, if it belongs to the user, they can edit, but if it's not theirs, they can just view it.

+++ login_screen/app/src/main/java/com/example/scratchappfeature/RecipeRVAdapter.java
@@ -18,14 +18,17 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.FieldValue;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
 import com.squareup.picasso.Picasso;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
 
 import classes.Recipe;
 
@@ -35,6 +38,7 @@ public class RecipeRVAdapter extends RecyclerView.Adapter<RecipeRVAdapter.Recipe
     private ArrayList<Recipe> recipeArrayListFull;
     private Context context;
     private OnItemClickListener mListener;
+    private String userID;
     private final FirebaseFirestore db = FirebaseFirestore.getInstance();
     private final FirebaseStorage storage = FirebaseStorage.getInstance();
     private final StorageReference storageReference = storage.getReference("images/");
@@ -46,7 +50,7 @@ public class RecipeRVAdapter extends RecyclerView.Adapter<RecipeRVAdapter.Recipe
     public void setOnItemClickListener(OnItemClickListener listener) {
         mListener = listener;
     }
-
+    public void setProfileID(String profileID){ this.userID = profileID;}
     public static class RecipeViewHolder extends RecyclerView.ViewHolder {
         // creating variables for our views of recycler items.
         private final TextView recipeNameTextView;
@@ -83,21 +87,30 @@ public class RecipeRVAdapter extends RecyclerView.Adapter<RecipeRVAdapter.Recipe
 
         Recipe recipeToDelete = recipeArrayList.get(position);
         String docID = recipeToDelete.getDocument_ID();
+        String userIDToDelete = recipeToDelete.getUser_ID();
 
-        db.collection("recipes").document(docID)
-            .delete()
-            .addOnSuccessListener(unused -> {
-                Log.i(TAG, "DocumentSnapshot successfully deleted!");
-                // Delete the image file as well after deleting the recipe
-                if(recipeToDelete.getImageName() != null) { // first check if the recipe has an image
-                    StorageReference recipeImageRef = storageReference.child(recipeToDelete.getImageName());
-                    Log.d(TAG, "Deleting image: " + recipeToDelete.getImageName());
-                    recipeImageRef.delete()
-                        .addOnSuccessListener(unused1 -> Log.i(TAG, "Successfully deleted image: " + recipeToDelete.getImageName()))
-                        .addOnFailureListener(e -> Log.e(TAG, e.toString()));
-                }
-            })
-            .addOnFailureListener(e -> Log.e(TAG, "Error deleting document", e));
+        if (!userIDToDelete.equals(userID)){
+            final Map<String, Object> savedRecipes = new HashMap<>();
+            savedRecipes.put("savedRecipes", FieldValue.arrayRemove(docID));
+            db.collection("profile")
+                    .document(userID)
+                    .update(savedRecipes);
+        } else {
+            db.collection("recipes").document(docID)
+                    .delete()
+                    .addOnSuccessListener(unused -> {
+                        Log.i(TAG, "DocumentSnapshot successfully deleted!");
+                        // Delete the image file as well after deleting the recipe
+                        if (recipeToDelete.getImageName() != null) { // first check if the recipe has an image
+                            StorageReference recipeImageRef = storageReference.child(recipeToDelete.getImageName());
+                            Log.d(TAG, "Deleting image: " + recipeToDelete.getImageName());
+                            recipeImageRef.delete()
+                                    .addOnSuccessListener(unused1 -> Log.i(TAG, "Successfully deleted image: " + recipeToDelete.getImageName()))
+                                    .addOnFailureListener(e -> Log.e(TAG, e.toString()));
+                        }
+                    })
+                    .addOnFailureListener(e -> Log.e(TAG, "Error deleting document", e));
+        }
         recipeArrayList.remove(position);
     }
